requirements.txt
dearpygui~=1.10.1
xdialog~=1.1.1.1
loguru~=0.7.2
pillow~=10.2.0